# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tu Tiên Lộ Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mission/{missionId}/accept:
    post:
      operationId: acceptMission
      tags: [mission]
      parameters:
        - name: missionId
          in: path
          required: true
          schema:
            type: string
      summary: Accept a mission
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /mission/{missionId}/complete:
    post:
      operationId: completeMission
      tags: [mission]
      parameters:
        - name: missionId
          in: path
          required: true
          schema:
            type: string
      summary: Complete and claim a mission
      responses:
        "200":
          description: Reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionReward"
...
  /sect/{sectId}/join:
    post:
      operationId: joinSect
      tags: [sect]
      parameters:
        - name: sectId
          in: path
          required: true
          schema:
            type: string
      summary: Join a sect
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /topup/request:
    post:
      operationId: requestTopup
      tags: [topup]
      summary: Submit a topup request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupRequestBody"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopupRequest"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    MissionReward:
      type: object
      properties:
        expGained:
          type: integer
        linhThachGained:
          type: integer
        itemsGranted:
          type: array
          items:
            type: string
        message:
          type: string
      required:
        - expGained
        - linhThachGained
        - itemsGranted
        - message
...
    TopupRequestBody:
      type: object
      properties:
        amount:
          type: integer
          minimum: 10000
          description: VND amount
        transferCode:
          type: string
      required:
        - amount
        - transferCode
    TopupRequest:
      type: object
      properties:
        id:
          type: string
        amount:
          type: integer
        tienNgocGranted:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        transferCode:
          type: string
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - amount
        - tienNgocGranted
        - status
        - transferCode
        - createdAt
...